refactor: modules structure
