+++ youtube-client-app/src/app/app.component.html
@@ -1,6 +1,2 @@
-<app-header
-  (searchResult)="showResult($event)"
-  (sortBlock)="showSortBlock($event)"
-></app-header>
-<div *ngIf="sortBlock" class="name"><app-sort></app-sort></div>
-<div *ngIf="result" class="name"><router-outlet></router-outlet></div>
+<header></header>
+<router-outlet></router-outlet>
