--- youtube-client-app/src/app/pages/main-page/main-page.component.html
@@ -1,3 +1,0 @@
-<div class="cards__container mt-14">
-  <app-card *ngFor="let card of cards" [card]="card"></app-card>
-</div>
