+++ youtube-client-app/src/app/shared/shared.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { ButtonComponent } from './components/button/button.component';
+
+@NgModule({
+  declarations: [ButtonComponent],
+  imports: [],
+  exports: [ButtonComponent],
+})
+export class SharedModule {}
