+++ youtube-client-app/src/app/youtube/components/card/card.component.ts
