+++ youtube-client-app/src/app/youtube/components/header/header.component.scss
