+++ youtube-client-app/src/app/youtube/components/search/search.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Output } from '@angular/core';
-import { SearchDataService } from 'src/app/services/search-data.service';
+import { SearchDataService } from 'src/app/youtube/services/search-data/search-data.service';
 
 @Component({
   selector: 'app-search',
