+++ youtube-client-app/src/app/youtube/directives/card-color/card-color.directive.ts
